Workaround/simplification of code in simple_cube.py.
This change makes simple_cube.py work in python 3.
It is necessary since the class scope is inaccessible from
functions, list comprehensions or generator expressions
enclosed in that scope; they act as if that scope does not exist.

see: https://stackoverflow.com/questions/13905741/accessing-class-variables-from-a-list-comprehension-in-the-class-definition

This code was tested and worked in 3.8.4 & 2.7.14